Extract identity wallet validator modules
