+++ lib-blockchain/src/blockchain.rs
@@ -30,9 +30,12 @@ mod test_utils;
 #[cfg(test)]
 mod tests;
 mod dao;
+mod identity;
 mod init;
 mod oracle;
 mod persistence;
+mod validators;
+mod wallets;
 
 pub use persistence::PersistenceStats;
 
@@ -2291,738 +2294,6 @@ impl Blockchain {
         }
     }
 
-    // ===== IDENTITY MANAGEMENT METHODS =====
-
-    /// Register a new identity on the blockchain
-    pub fn register_identity(&mut self, identity_data: IdentityTransactionData) -> Result<Hash> {
-        // Check if identity already exists
-        if self.identity_registry.contains_key(&identity_data.did) {
-            return Err(anyhow::anyhow!(
-                "Identity {} already exists on blockchain",
-                identity_data.did
-            ));
-        }
-
-        // Create identity registration transaction
-        let registration_tx = Transaction::new_identity_registration(
-            identity_data.clone(),
-            vec![], // Fee outputs handled separately
-            Signature {
-                signature: identity_data.ownership_proof.clone(),
-                public_key: PublicKey::new(identity_data.public_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: identity_data.created_at,
-            },
-            format!("Identity registration for {}", identity_data.did).into_bytes(),
-        );
-
-        // Add to pending transactions for inclusion in next block
-        self.add_pending_transaction(registration_tx.clone())?;
-
-        // Store in identity registry immediately for queries
-        self.identity_registry
-            .insert(identity_data.did.clone(), identity_data.clone());
-        self.identity_blocks
-            .insert(identity_data.did.clone(), self.height + 1);
-
-        Ok(registration_tx.hash())
-    }
-
-    /// Register a new identity on the blockchain with persistent storage
-    pub async fn register_identity_with_persistence(
-        &mut self,
-        identity_data: IdentityTransactionData,
-    ) -> Result<Hash> {
-        // Register identity normally
-        let tx_hash = self.register_identity(identity_data.clone())?;
-
-        // Store identity data in persistent storage if available
-        if let Some(storage_manager_arc) = &self.storage_manager {
-            let mut storage_manager = storage_manager_arc.write().await;
-            if let Err(e) = storage_manager
-                .store_identity_data(&identity_data.did, &identity_data)
-                .await
-            {
-                eprintln!("Warning: Failed to persist identity data to storage: {}", e);
-            }
-        }
-
-        Ok(tx_hash)
-    }
-
-    /// Get identity data from blockchain
-    pub fn get_identity(&self, did: &str) -> Option<&IdentityTransactionData> {
-        self.identity_registry.get(did)
-    }
-
-    /// Check if identity exists on blockchain
-    pub fn identity_exists(&self, did: &str) -> bool {
-        self.identity_registry.contains_key(did)
-    }
-
-    /// Update an existing identity on the blockchain
-    pub fn update_identity(
-        &mut self,
-        did: &str,
-        updated_data: IdentityTransactionData,
-    ) -> Result<Hash> {
-        // Check if identity exists
-        let existing = self
-            .identity_registry
-            .get(did)
-            .ok_or_else(|| anyhow::anyhow!("Identity {} not found on blockchain", did))?;
-
-        // Enforce immutable ownership/identity invariants
-        if existing.did != updated_data.did {
-            return Err(anyhow::anyhow!(
-                "Immutable DID mismatch for identity update"
-            ));
-        }
-        if existing.public_key != updated_data.public_key {
-            return Err(anyhow::anyhow!(
-                "Immutable public key mismatch for identity update"
-            ));
-        }
-        if existing.identity_type != updated_data.identity_type {
-            return Err(anyhow::anyhow!(
-                "Immutable identity type mismatch for identity update"
-            ));
-        }
-
-        // Create update transaction with authorization
-        let auth_input = TransactionInput {
-            previous_output: Hash::default(),
-            output_index: 0,
-            nullifier: crate::types::hash::blake3_hash(
-                &format!("identity_update_{}", did).as_bytes(),
-            ),
-            zk_proof: ZkTransactionProof::default(),
-        };
-
-        let update_tx = Transaction::new_identity_update(
-            updated_data.clone(),
-            vec![auth_input],
-            vec![], // No outputs needed
-            100,    // Update fee
-            Signature {
-                signature: updated_data.ownership_proof.clone(),
-                public_key: PublicKey::new(updated_data.public_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: updated_data.created_at,
-            },
-            format!("Identity update for {}", did).into_bytes(),
-        );
-
-        // Add to pending transactions
-        self.add_pending_transaction(update_tx.clone())?;
-
-        // Update registry
-        self.identity_registry.insert(did.to_string(), updated_data);
-
-        Ok(update_tx.hash())
-    }
-
-    /// Update an existing identity on the blockchain with persistent storage
-    pub async fn update_identity_with_persistence(
-        &mut self,
-        did: &str,
-        updated_data: IdentityTransactionData,
-    ) -> Result<Hash> {
-        // Update identity normally
-        let tx_hash = self.update_identity(did, updated_data.clone())?;
-
-        // Store updated identity data in persistent storage if available
-        if let Some(storage_manager_arc) = &self.storage_manager {
-            let mut storage_manager = storage_manager_arc.write().await;
-            if let Err(e) = storage_manager
-                .store_identity_data(did, &updated_data)
-                .await
-            {
-                eprintln!(
-                    "Warning: Failed to persist updated identity data to storage: {}",
-                    e
-                );
-            }
-        }
-
-        Ok(tx_hash)
-    }
-
-    /// Revoke an identity on the blockchain
-    pub fn revoke_identity(&mut self, did: &str, authorizing_signature: Vec<u8>) -> Result<Hash> {
-        // Check if identity exists
-        if !self.identity_registry.contains_key(did) {
-            return Err(anyhow::anyhow!("Identity {} not found on blockchain", did));
-        }
-
-        // Create authorization input from existing identity
-        let auth_input = TransactionInput {
-            previous_output: Hash::default(),
-            output_index: 0,
-            nullifier: crate::types::hash::blake3_hash(
-                &format!("identity_revoke_{}", did).as_bytes(),
-            ),
-            zk_proof: ZkTransactionProof::default(),
-        };
-
-        let revocation_tx = Transaction::new_identity_revocation(
-            did.to_string(),
-            vec![auth_input],
-            50, // Revocation fee
-            Signature {
-                signature: authorizing_signature,
-                public_key: PublicKey::new(vec![]), // Would be filled from auth
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: crate::utils::time::current_timestamp(),
-            },
-            format!("Identity revocation for {}", did).into_bytes(),
-        );
-
-        // Add to pending transactions
-        self.add_pending_transaction(revocation_tx.clone())?;
-
-        // Remove from registry (mark as revoked)
-        if let Some(mut identity_data) = self.identity_registry.remove(did) {
-            identity_data.identity_type = "revoked".to_string();
-            self.identity_registry
-                .insert(format!("{}_revoked", did), identity_data);
-        }
-
-        Ok(revocation_tx.hash())
-    }
-
-    /// Get all identities on the blockchain
-    pub fn list_all_identities(&self) -> Vec<&IdentityTransactionData> {
-        self.identity_registry.values().collect()
-    }
-
-    /// Get all identities as HashMap
-    pub fn get_all_identities(&self) -> &HashMap<String, IdentityTransactionData> {
-        &self.identity_registry
-    }
-
-    /// Get identity block confirmation count
-    pub fn get_identity_confirmations(&self, did: &str) -> Option<u64> {
-        self.identity_blocks.get(did).map(|block_height| {
-            if self.height >= *block_height {
-                self.height - block_height + 1
-            } else {
-                0
-            }
-        })
-    }
-
-    /// Process identity transactions in a block
-    ///
-    /// This method:
-    /// 1. Stores to in-memory HashMap (for fast queries, backward compatibility)
-    /// 2. Persists to sled storage (for durability, consensus state)
-    pub fn process_identity_transactions(&mut self, block: &Block) -> Result<()> {
-        for transaction in &block.transactions {
-            if transaction.transaction_type.is_identity_transaction() {
-                if let Some(identity_data) = transaction.identity_data() {
-                    match transaction.transaction_type {
-                        TransactionType::IdentityRegistration => {
-                            // CRITICAL: Preserve controlled_nodes if identity already exists
-                            let mut new_identity_data = identity_data.clone();
-                            if let Some(existing_identity) =
-                                self.identity_registry.get(&identity_data.did)
-                            {
-                                // Preserve controlled_nodes from existing identity
-                                new_identity_data.controlled_nodes =
-                                    existing_identity.controlled_nodes.clone();
-                            }
-
-                            // Store in memory (backward compatibility + fast queries)
-                            self.identity_registry
-                                .insert(identity_data.did.clone(), new_identity_data.clone());
-                            self.identity_blocks
-                                .insert(identity_data.did.clone(), block.height());
-
-                            // PHASE 0: Persist to sled storage (consensus state)
-                            if let Some(ref store) = self.store {
-                                self.persist_identity_registration(
-                                    store.as_ref(),
-                                    &new_identity_data,
-                                    block.height(),
-                                )?;
-                            }
-
-                            // Register for UBI if this is a citizen identity
-                            if identity_data.identity_type == "verified_citizen"
-                                || identity_data.identity_type == "citizen"
-                                || identity_data.identity_type == "external_citizen"
-                            {
-                                // Find the UBI wallet from owned_wallets
-                                let ubi_wallet_id = new_identity_data
-                                    .owned_wallets
-                                    .iter()
-                                    .find(|wallet_id| {
-                                        self.wallet_registry
-                                            .get(*wallet_id)
-                                            .map(|w| w.wallet_type == "UBI")
-                                            .unwrap_or(false)
-                                    })
-                                    .cloned();
-
-                                if let Some(ubi_wallet) = ubi_wallet_id {
-                                    if let Err(e) = self.register_for_ubi(
-                                        identity_data.did.clone(),
-                                        ubi_wallet,
-                                        block.height(),
-                                    ) {
-                                        warn!(
-                                            "Failed to register {} for UBI: {}",
-                                            identity_data.did, e
-                                        );
-                                    }
-                                } else {
-                                    warn!("No UBI wallet found for citizen {}", identity_data.did);
-                                }
-                            }
-                        }
-                        TransactionType::IdentityUpdate => {
-                            // CRITICAL: Preserve controlled_nodes on update
-                            let mut updated_identity_data = identity_data.clone();
-                            if let Some(existing_identity) =
-                                self.identity_registry.get(&identity_data.did)
-                            {
-                                // Preserve controlled_nodes from existing identity
-                                updated_identity_data.controlled_nodes =
-                                    existing_identity.controlled_nodes.clone();
-                            }
-
-                            // Enforce immutable ownership and identity invariants
-                            if let Some(existing_identity) =
-                                self.identity_registry.get(&identity_data.did)
-                            {
-                                if existing_identity.public_key != updated_identity_data.public_key
-                                {
-                                    return Err(anyhow::anyhow!(
-                                        "Immutable public key mismatch for identity update: {}",
-                                        identity_data.did
-                                    ));
-                                }
-                                if existing_identity.identity_type
-                                    != updated_identity_data.identity_type
-                                {
-                                    return Err(anyhow::anyhow!(
-                                        "Immutable identity type mismatch for identity update: {}",
-                                        identity_data.did
-                                    ));
-                                }
-                            } else {
-                                return Err(anyhow::anyhow!(
-                                    "Cannot update non-existent identity: {}",
-                                    identity_data.did
-                                ));
-                            }
-
-                            // Store in memory (post-validation)
-                            self.identity_registry
-                                .insert(identity_data.did.clone(), updated_identity_data.clone());
-
-                            // PHASE 0: Persist update to sled storage
-                            if let Some(ref store) = self.store {
-                                self.persist_identity_update(
-                                    store.as_ref(),
-                                    &updated_identity_data,
-                                )?;
-                            }
-                        }
-                        TransactionType::IdentityRevocation => {
-                            let did_hash = did_to_hash(&identity_data.did);
-
-                            // Store revoked state in memory
-                            let mut revoked_data = identity_data.clone();
-                            revoked_data.identity_type = "revoked".to_string();
-                            self.identity_registry
-                                .insert(format!("{}_revoked", identity_data.did), revoked_data);
-                            self.identity_registry.remove(&identity_data.did);
-
-                            // PHASE 0: Delete from sled storage (identity + indexes)
-                            if let Some(ref store) = self.store {
-                                if let Some(existing_identity) =
-                                    store.get_identity(&did_hash).map_err(|e| {
-                                        anyhow::anyhow!(
-                                            "Failed to load identity for revocation: {}",
-                                            e
-                                        )
-                                    })?
-                                {
-                                    store
-                                        .delete_identity_owner_index(&existing_identity.owner)
-                                        .map_err(|e| {
-                                            anyhow::anyhow!(
-                                                "Failed to delete identity owner index: {}",
-                                                e
-                                            )
-                                        })?;
-                                }
-                                store.delete_identity(&did_hash).map_err(|e| {
-                                    anyhow::anyhow!("Failed to delete identity from sled: {}", e)
-                                })?;
-                                store.delete_identity_metadata(&did_hash).map_err(|e| {
-                                    anyhow::anyhow!(
-                                        "Failed to delete identity metadata from sled: {}",
-                                        e
-                                    )
-                                })?;
-                            }
-                        }
-                        _ => {} // Other transaction types
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Persist a newly-registered identity to sled (registration only).
-    fn persist_identity_registration(
-        &self,
-        store: &dyn BlockchainStore,
-        identity_data: &IdentityTransactionData,
-        block_height: u64,
-    ) -> Result<()> {
-        use crate::storage::derive_address_from_public_key;
-        use crate::types::hash::blake3_hash;
-
-        let did_hash = did_to_hash(&identity_data.did);
-
-        // Derive owner address from public key using the canonical helper
-        let owner = derive_address_from_public_key(&identity_data.public_key);
-
-        // Convert to consensus-compliant fixed-size format
-        let consensus = IdentityConsensus {
-            did_hash,
-            owner,
-            public_key_hash: blake3_hash(&identity_data.public_key).as_array(),
-            did_document_hash: identity_data.did_document_hash.as_array(),
-            seed_commitment: None, // Will be set during migration if available
-            identity_type: IdentityType::from_str(&identity_data.identity_type),
-            status: IdentityStatus::Active,
-            version: 1, // Legacy format
-            created_at: identity_data.created_at,
-            registered_at_height: block_height,
-            registration_fee: identity_data.registration_fee,
-            dao_fee: identity_data.dao_fee,
-            controlled_node_count: identity_data.controlled_nodes.len() as u32,
-            owned_wallet_count: identity_data.owned_wallets.len() as u32,
-            attribute_count: 0,
-        };
-
-        // Convert to metadata (allows strings)
-        let metadata = IdentityMetadata {
-            did: identity_data.did.clone(),
-            display_name: identity_data.display_name.clone(),
-            public_key: identity_data.public_key.clone(),
-            ownership_proof: identity_data.ownership_proof.clone(),
-            controlled_nodes: identity_data.controlled_nodes.clone(),
-            owned_wallets: identity_data.owned_wallets.clone(),
-            attributes: Vec::new(),
-        };
-
-        // Persist to sled
-        store
-            .put_identity(&did_hash, &consensus)
-            .map_err(|e| anyhow::anyhow!("Failed to store identity in sled: {}", e))?;
-        store
-            .put_identity_metadata(&did_hash, &metadata)
-            .map_err(|e| anyhow::anyhow!("Failed to store identity metadata in sled: {}", e))?;
-        store
-            .put_identity_owner_index(&consensus.owner, &did_hash)
-            .map_err(|e| anyhow::anyhow!("Failed to store identity owner index in sled: {}", e))?;
-
-        debug!(
-            "Persisted identity {} to sled storage (registration)",
-            identity_data.did
-        );
-        Ok(())
-    }
-
-    /// Persist an identity update to sled (update only).
-    fn persist_identity_update(
-        &self,
-        store: &dyn BlockchainStore,
-        identity_data: &IdentityTransactionData,
-    ) -> Result<()> {
-        use crate::storage::derive_address_from_public_key;
-        use crate::types::hash::blake3_hash;
-
-        let did_hash = did_to_hash(&identity_data.did);
-
-        let existing = store
-            .get_identity(&did_hash)
-            .map_err(|e| anyhow::anyhow!("Failed to load identity for update: {}", e))?
-            .ok_or_else(|| {
-                anyhow::anyhow!("Cannot update non-existent identity: {}", identity_data.did)
-            })?;
-
-        let existing_metadata = store
-            .get_identity_metadata(&did_hash)
-            .map_err(|e| anyhow::anyhow!("Failed to load identity metadata for update: {}", e))?
-            .ok_or_else(|| {
-                anyhow::anyhow!(
-                    "Missing identity metadata for update: {}",
-                    identity_data.did
-                )
-            })?;
-
-        // Enforce immutable ownership and identity invariants
-        let incoming_owner = derive_address_from_public_key(&identity_data.public_key);
-        let incoming_public_key_hash = blake3_hash(&identity_data.public_key).as_array();
-        let incoming_identity_type = IdentityType::from_str(&identity_data.identity_type);
-
-        if existing.did_hash != did_hash {
-            return Err(anyhow::anyhow!(
-                "Immutable DID hash mismatch for identity update"
-            ));
-        }
-        if existing.owner != incoming_owner {
-            return Err(anyhow::anyhow!(
-                "Immutable owner mismatch for identity update"
-            ));
-        }
-        if existing.public_key_hash != incoming_public_key_hash {
-            return Err(anyhow::anyhow!(
-                "Immutable public key mismatch for identity update"
-            ));
-        }
-        if existing.identity_type != incoming_identity_type {
-            return Err(anyhow::anyhow!(
-                "Immutable identity type mismatch for identity update"
-            ));
-        }
-        if existing_metadata.did != identity_data.did {
-            return Err(anyhow::anyhow!(
-                "Immutable DID mismatch for identity update"
-            ));
-        }
-        if existing_metadata.public_key != identity_data.public_key {
-            return Err(anyhow::anyhow!(
-                "Immutable public key mismatch for identity update"
-            ));
-        }
-
-        // Apply validated diff: consensus keeps immutable fields, update mutable fields only
-        let mut updated_consensus = existing.clone();
-        updated_consensus.did_document_hash = identity_data.did_document_hash.as_array();
-        updated_consensus.controlled_node_count = identity_data.controlled_nodes.len() as u32;
-        updated_consensus.owned_wallet_count = identity_data.owned_wallets.len() as u32;
-
-        let mut updated_metadata = existing_metadata.clone();
-        updated_metadata.display_name = identity_data.display_name.clone();
-        updated_metadata.ownership_proof = identity_data.ownership_proof.clone();
-        updated_metadata.controlled_nodes = identity_data.controlled_nodes.clone();
-        updated_metadata.owned_wallets = identity_data.owned_wallets.clone();
-
-        store
-            .put_identity(&did_hash, &updated_consensus)
-            .map_err(|e| anyhow::anyhow!("Failed to update identity in sled: {}", e))?;
-        store
-            .put_identity_metadata(&did_hash, &updated_metadata)
-            .map_err(|e| anyhow::anyhow!("Failed to update identity metadata in sled: {}", e))?;
-
-        debug!(
-            "Persisted identity {} to sled storage (update)",
-            identity_data.did
-        );
-        Ok(())
-    }
-
-    /// Check if a public key is registered as an identity on the blockchain
-    pub fn is_public_key_registered(&self, public_key: &[u8]) -> bool {
-        for identity_data in self.identity_registry.values() {
-            if identity_data.public_key == public_key && identity_data.identity_type != "revoked" {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Get identity by public key
-    pub fn get_identity_by_public_key(
-        &self,
-        public_key: &[u8],
-    ) -> Option<&IdentityTransactionData> {
-        for identity_data in self.identity_registry.values() {
-            if identity_data.public_key == public_key && identity_data.identity_type != "revoked" {
-                return Some(identity_data);
-            }
-        }
-        None
-    }
-
-    /// Auto-register wallet identity if not already registered (system transaction)
-    /// This creates a minimal identity registration for wallets that don't have one
-    pub fn auto_register_wallet_identity(
-        &mut self,
-        wallet_id: &str,
-        public_key: Vec<u8>,
-        did: Option<String>,
-    ) -> Result<Hash> {
-        // Check if this public key is already registered
-        if self.is_public_key_registered(&public_key) {
-            tracing::info!(" Public key already registered on blockchain");
-            return Ok(Hash::default());
-        }
-
-        // Generate DID from wallet ID if not provided
-        let identity_did =
-            did.unwrap_or_else(|| format!("did:zhtp:wallet-{}", hex::encode(&public_key[..16])));
-
-        tracing::info!(" Auto-registering wallet identity: {}", identity_did);
-
-        // Create identity transaction data
-        let identity_data = IdentityTransactionData {
-            did: identity_did.clone(),
-            display_name: format!("Wallet {}", &wallet_id[..8.min(wallet_id.len())]),
-            public_key: public_key.clone(),
-            ownership_proof: vec![], // System transaction doesn't need proof
-            identity_type: "service".to_string(), // Use "service" type for wallet identities
-            did_document_hash: crate::types::hash::blake3_hash(identity_did.as_bytes()),
-            created_at: crate::utils::time::current_timestamp(),
-            registration_fee: 0, // No fee for auto-registration
-            dao_fee: 0,
-            controlled_nodes: Vec::new(),
-            owned_wallets: vec![wallet_id.to_string()],
-        };
-
-        // Create identity registration transaction as system transaction
-        let registration_tx = Transaction::new_identity_registration(
-            identity_data.clone(),
-            vec![], // No outputs for system transaction
-            Signature {
-                signature: vec![0xAA; 64], // System signature marker
-                public_key: PublicKey::new(public_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: identity_data.created_at,
-            },
-            b"Auto-registration for wallet identity".to_vec(),
-        );
-
-        // Add as system transaction (bypasses normal validation)
-        self.add_system_transaction(registration_tx.clone())?;
-
-        // Register in identity registry immediately
-        self.identity_registry
-            .insert(identity_did.clone(), identity_data.clone());
-        self.identity_blocks.insert(identity_did, self.height + 1);
-
-        tracing::info!(" Wallet identity auto-registered on blockchain");
-
-        Ok(registration_tx.hash())
-    }
-
-    /// Ensure wallet identity is registered before transaction (convenience method)
-    pub fn ensure_wallet_identity_registered(
-        &mut self,
-        wallet_id: &str,
-        public_key: &[u8],
-        did: Option<String>,
-    ) -> Result<()> {
-        if !self.is_public_key_registered(public_key) {
-            self.auto_register_wallet_identity(wallet_id, public_key.to_vec(), did)?;
-        }
-        Ok(())
-    }
-
-    // ===== WALLET MANAGEMENT METHODS =====
-
-    /// Ensure the native SOV token contract exists in memory.
-    fn ensure_sov_token_contract(&mut self) {
-        let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-        if !self.token_contracts.contains_key(&sov_token_id) {
-            let sov_token = crate::contracts::TokenContract::new_sov_native();
-            self.token_contracts.insert(sov_token_id, sov_token);
-            info!("🪙 Initialized native SOV token contract");
-        }
-    }
-
-    /// Ensure the deterministic DAO treasury wallet exists in the registry.
-    ///
-    /// Uses `blake3(b"SOV_DAO_TREASURY_V1")` as the wallet ID so every node
-    /// derives the same identity independently.  Idempotent: a second call is a
-    /// no-op when the wallet is already present and linked.
-    fn ensure_treasury_wallet(&mut self) {
-        // Deterministic ID — identical on every node.
-        let wallet_id_bytes = Self::deterministic_treasury_wallet_id().as_array();
-        let wallet_id_hex = hex::encode(wallet_id_bytes);
-
-        // Insert into registry if not present.
-        if !self.wallet_registry.contains_key(&wallet_id_hex) {
-            let wallet_data = crate::transaction::WalletTransactionData {
-                wallet_id: crate::types::Hash::new(wallet_id_bytes),
-                wallet_type: "treasury".to_string(),
-                wallet_name: "DAO Treasury".to_string(),
-                alias: None,
-                // public_key is intentionally empty: the DAO treasury wallet uses the
-                // balance model (token.balances keyed by wallet_key_for_sov(wallet_id)),
-                // not the UTXO model. UTXO-based treasury paths (get_dao_treasury_utxos,
-                // execute_dao_proposal) are legacy and do not apply to this wallet.
-                public_key: vec![],
-                owner_identity_id: None,
-                seed_commitment: crate::types::Hash::zero(),
-                created_at: 0,
-                registration_fee: 0,
-                capabilities: 0,
-                initial_balance: 0,
-            };
-            self.wallet_registry
-                .insert(wallet_id_hex.clone(), wallet_data);
-        }
-
-        // Link as the active treasury wallet if not already set.
-        if self.dao_treasury_wallet_id.is_none() {
-            self.dao_treasury_wallet_id = Some(wallet_id_hex);
-            info!("🏦 DAO treasury wallet initialized (deterministic bootstrap)");
-        }
-    }
-
-    /// Deterministic DAO treasury wallet ID shared by every node.
-    pub fn deterministic_treasury_wallet_id() -> crate::types::Hash {
-        crate::types::hash::blake3_hash(b"SOV_DAO_TREASURY_V1")
-    }
-
-    /// Returns true if a wallet exists in canonical history, either via genesis state
-    /// (`wallet_blocks[id] = 0`) or via a committed WalletRegistration transaction.
-    pub fn wallet_exists_in_canonical_history(&self, wallet_id: &crate::types::Hash) -> bool {
-        let wallet_id_hex = hex::encode(wallet_id.as_bytes());
-        if self.wallet_blocks.get(&wallet_id_hex).copied() == Some(0) {
-            return true;
-        }
-
-        self.blocks.iter().any(|block| {
-            block.transactions.iter().any(|tx| {
-                tx.wallet_data()
-                    .map(|wallet_data| wallet_data.wallet_id == *wallet_id)
-                    .unwrap_or(false)
-            })
-        })
-    }
-
-    /// Return wallets present in local state but absent from committed block history.
-    pub fn collect_noncanonical_wallets(&self) -> Vec<crate::transaction::WalletTransactionData> {
-        let mut wallets: Vec<_> = self
-            .wallet_registry
-            .values()
-            .filter(|wallet| !self.wallet_exists_in_canonical_history(&wallet.wallet_id))
-            .cloned()
-            .collect();
-        wallets.sort_by_key(|wallet| hex::encode(wallet.wallet_id.as_bytes()));
-        wallets
-    }
-
-    /// Returns true when the configured DAO treasury wallet exists in committed block history.
-    pub fn dao_treasury_wallet_is_canonical(&self) -> bool {
-        self.dao_treasury_wallet_id
-            .as_ref()
-            .and_then(|wallet_id_hex| self.wallet_registry.get(wallet_id_hex))
-            .map(|wallet| self.wallet_exists_in_canonical_history(&wallet.wallet_id))
-            .unwrap_or(false)
-    }
-
     /// Apply a token transfer with protocol fee deduction and treasury routing.
     ///
     /// This is a helper that consolidates the duplicated fee logic from the two
@@ -3092,241 +2363,6 @@ impl Blockchain {
         }
         evicted
     }
-    fn resolve_credit_pubkey_from_parts(
-        &self,
-        public_key: Vec<u8>,
-        owner_identity_id: Option<Hash>,
-    ) -> Option<Vec<u8>> {
-        if public_key.len() >= Self::MIN_DILITHIUM_PK_LEN {
-            return Some(public_key);
-        }
-
-        if let Some(owner) = owner_identity_id {
-            let did = format!("did:zhtp:{}", hex::encode(owner.as_bytes()));
-            if let Some(identity) = self.identity_registry.get(&did) {
-                if identity.public_key.len() >= Self::MIN_DILITHIUM_PK_LEN {
-                    return Some(identity.public_key.clone());
-                }
-            }
-        }
-
-        warn!(
-            "SOV credit skipped: short public key (len={}) and no full identity key",
-            public_key.len()
-        );
-        None
-    }
-
-    fn resolve_wallet_credit_pubkey(
-        &self,
-        wallet: &crate::transaction::WalletTransactionData,
-    ) -> Option<Vec<u8>> {
-        self.resolve_credit_pubkey_from_parts(
-            wallet.public_key.clone(),
-            wallet.owner_identity_id.clone(),
-        )
-    }
-
-    /// Check whether a token_id refers to native SOV (zero or legacy SOV token id).
-    fn is_sov_token_id(token_id: &[u8; 32]) -> bool {
-        *token_id == [0u8; 32] || *token_id == crate::contracts::utils::generate_lib_token_id()
-    }
-
-    /// Create a synthetic PublicKey keyed by wallet_id for SOV balances.
-    /// This uses an empty keypair and the wallet_id bytes as key_id.
-    fn wallet_key_for_sov(wallet_id: &[u8; 32]) -> PublicKey {
-        crate::contracts::utils::wallet_key_for_sov(*wallet_id)
-    }
-
-    /// Initialize Treasury Kernel with SOV token authority.
-    /// Must be called after SOV token is created with kernel authority.
-    pub fn initialize_treasury_kernel(&mut self, kernel_authority: PublicKey) {
-        use crate::contracts::treasury_kernel::TreasuryKernel;
-
-        let governance_authority = kernel_authority.clone();
-        self.treasury_kernel = Some(TreasuryKernel::new(
-            kernel_authority,
-            governance_authority,
-            100, // blocks per epoch
-        ));
-        info!("Treasury Kernel initialized");
-    }
-
-    /// Check if a token is controlled by Treasury Kernel.
-    /// Returns true if token has kernel_mint_authority set.
-    fn is_kernel_controlled_token(&self, token: &crate::contracts::TokenContract) -> bool {
-        token.kernel_mint_authority.is_some()
-    }
-
-    /// Credit tokens to an account - routes through Treasury Kernel for SOV/DAO tokens,
-    /// uses direct method for custom tokens.
-    ///
-    /// SECURITY: For kernel-controlled tokens, this REQUIRES the kernel to be initialized.
-    /// There is NO fallback to direct methods for security reasons.
-    fn credit_tokens(
-        &mut self,
-        token: &mut crate::contracts::TokenContract,
-        to: &PublicKey,
-        amount: u64,
-        reason: crate::contracts::treasury_kernel::CreditReason,
-    ) -> Result<(), String> {
-        // Check if token is kernel-controlled (SOV, DAO tokens)
-        if self.is_kernel_controlled_token(token) {
-            // Must route through Treasury Kernel - no fallback for security
-            let kernel = self.treasury_kernel.as_mut()
-                .ok_or_else(|| "Treasury Kernel not initialized - kernel-controlled token operations require kernel".to_string())?;
-            let caller = kernel.governance_authority().clone();
-            kernel
-                .credit(token, &caller, to, amount, reason)
-                .map_err(|e| e.to_string())
-        } else {
-            // Custom token - use direct method (no kernel control)
-            token.credit_balance(to, amount)
-        }
-    }
-
-    /// Debit tokens from an account - routes through Treasury Kernel for SOV/DAO tokens,
-    /// uses direct method for custom tokens.
-    ///
-    /// SECURITY: For kernel-controlled tokens, this REQUIRES the kernel to be initialized.
-    /// There is NO fallback to direct methods for security reasons.
-    fn debit_tokens(
-        &mut self,
-        token: &mut crate::contracts::TokenContract,
-        from: &PublicKey,
-        amount: u64,
-        reason: crate::contracts::treasury_kernel::DebitReason,
-    ) -> Result<(), String> {
-        // Check if token is kernel-controlled (SOV, DAO tokens)
-        if self.is_kernel_controlled_token(token) {
-            // Must route through Treasury Kernel - no fallback for security
-            let kernel = self.treasury_kernel.as_mut()
-                .ok_or_else(|| "Treasury Kernel not initialized - kernel-controlled token operations require kernel".to_string())?;
-            let caller = kernel.governance_authority().clone();
-            kernel
-                .debit(token, &caller, from, amount, reason)
-                .map_err(|e| e.to_string())
-        } else {
-            // Custom token - use direct method (no kernel control)
-            token.debit_balance(from, amount)
-        }
-    }
-
-    /// Convert wallet_id hex string to a 32-byte array.
-    fn wallet_id_bytes(wallet_id_hex: &str) -> Option<[u8; 32]> {
-        let bytes = hex::decode(wallet_id_hex).ok()?;
-        if bytes.len() != 32 {
-            return None;
-        }
-        let mut arr = [0u8; 32];
-        arr.copy_from_slice(&bytes);
-        Some(arr)
-    }
-
-    /// Find the Primary wallet_id for a signer key_id, if available.
-    pub fn primary_wallet_for_signer(&self, signer_key_id: &[u8; 32]) -> Option<[u8; 32]> {
-        for (wallet_id, wallet) in &self.wallet_registry {
-            if wallet.wallet_type != "Primary" {
-                continue;
-            }
-            let pk = PublicKey::new(wallet.public_key.clone());
-            if &pk.key_id == signer_key_id {
-                return Self::wallet_id_bytes(wallet_id);
-            }
-        }
-        None
-    }
-
-    /// Public accessor: find the Primary wallet ID bytes for a given signer key_id.
-    pub fn primary_wallet_id_for_signer(&self, signer_key_id: &[u8; 32]) -> Option<[u8; 32]> {
-        self.primary_wallet_for_signer(signer_key_id)
-    }
-
-    /// Public accessor: build the SOV-ledger lookup key for a wallet ID.
-    pub fn sov_key_from_wallet_id(wallet_id: &[u8; 32]) -> PublicKey {
-        Self::wallet_key_for_sov(wallet_id)
-    }
-
-    /// Migrate legacy SOV balances keyed by public-key key_id into Primary wallet_id entries.
-    fn migrate_sov_key_balances_to_wallets(&mut self) {
-        let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-        let token = match self.token_contracts.get_mut(&sov_token_id) {
-            Some(token) => token,
-            None => return,
-        };
-
-        // Map signer key_id -> primary wallet_id
-        let mut key_to_wallet: std::collections::HashMap<[u8; 32], [u8; 32]> =
-            std::collections::HashMap::new();
-        for (wallet_id, wallet) in &self.wallet_registry {
-            if wallet.wallet_type != "Primary" {
-                continue;
-            }
-            if let Some(wallet_id_bytes) = Self::wallet_id_bytes(wallet_id) {
-                let pk = PublicKey::new(wallet.public_key.clone());
-                key_to_wallet.insert(pk.key_id, wallet_id_bytes);
-            }
-        }
-
-        let mut migrated_total: u64 = 0;
-        let balances: Vec<(PublicKey, u64)> = token
-            .balances
-            .iter()
-            .map(|(k, v)| (k.clone(), *v))
-            .collect();
-        for (pk, bal) in balances {
-            if bal == 0 {
-                continue;
-            }
-            // Skip entries already keyed by wallet_id.
-            let key_hex = hex::encode(pk.key_id);
-            if self.wallet_registry.contains_key(&key_hex) {
-                continue;
-            }
-            if let Some(wallet_id_bytes) = key_to_wallet.get(&pk.key_id) {
-                token.balances.remove(&pk);
-                let wallet_key = Self::wallet_key_for_sov(wallet_id_bytes);
-                let existing = token.balance_of(&wallet_key);
-                token
-                    .balances
-                    .insert(wallet_key, existing.saturating_add(bal));
-                migrated_total = migrated_total.saturating_add(bal);
-            }
-        }
-
-        if migrated_total > 0 {
-            info!(
-                "🪙 Migrated {} SOV from key-based balances to Primary wallets",
-                migrated_total
-            );
-        }
-    }
-
-    /// Resolve a full public key from a key_id by searching wallet and identity registries.
-    fn resolve_public_key_by_key_id(&self, key_id: &[u8; 32]) -> Option<Vec<u8>> {
-        for wallet in self.wallet_registry.values() {
-            if wallet.public_key.is_empty() {
-                continue;
-            }
-            let pk = PublicKey::new(wallet.public_key.clone());
-            if &pk.key_id == key_id {
-                return Some(wallet.public_key.clone());
-            }
-        }
-
-        for identity in self.identity_registry.values() {
-            if identity.public_key.is_empty() {
-                continue;
-            }
-            let pk = PublicKey::new(identity.public_key.clone());
-            if &pk.key_id == key_id {
-                return Some(identity.public_key.clone());
-            }
-        }
-
-        None
-    }
-
     /// Get the current expected nonce for a sender address and token.
     /// For SOV transfers, the address is the wallet_id bytes.
     /// For custom token transfers, the address is the key_id bytes.
@@ -3350,408 +2386,6 @@ impl Blockchain {
             }
         }
         0
-    }
-
-    /// Collect SOV backfill entries from wallet_registry for wallets missing token balances.
-    ///
-    /// Returns a list of (public_key_bytes, amount, wallet_id) that should be minted
-    /// via TokenMint transactions in a migration block.
-    pub fn collect_sov_backfill_entries(&self) -> Vec<([u8; 32], u64, String)> {
-        let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-        let token_opt = self.token_contracts.get(&sov_token_id);
-
-        let mut entries: Vec<([u8; 32], u64, String)> = Vec::new();
-        for (wallet_id, wallet) in &self.wallet_registry {
-            // Only backfill wallets that are already registered on-chain.
-            // This prevents minting to wallets that only exist in local state.
-            let is_on_chain = self
-                .wallet_blocks
-                .get(wallet_id)
-                .map(|h| *h <= self.height)
-                .unwrap_or(false);
-            if !is_on_chain {
-                continue;
-            }
-            if wallet.initial_balance == 0 {
-                continue;
-            }
-            let wallet_key = match Self::wallet_id_bytes(wallet_id) {
-                Some(bytes) => bytes,
-                None => {
-                    warn!(
-                        "Skipping SOV backfill for wallet {}: invalid wallet_id",
-                        &wallet_id[..16.min(wallet_id.len())]
-                    );
-                    continue;
-                }
-            };
-
-            let recipient = Self::wallet_key_for_sov(&wallet_key);
-            // Prefer Sled token_balances tree (authoritative when executor is active)
-            // over in-memory token_contracts.balances, which is never updated after
-            // executor-path TokenMint transactions. Using stale in-memory balances
-            // causes repeat minting on every restart.
-            let current_balance: u64 = if let Some(store) = self.get_store() {
-                let sov_storage_token_id = crate::storage::TokenId(sov_token_id);
-                let addr = crate::storage::Address::new(wallet_key);
-                store
-                    .get_token_balance(&sov_storage_token_id, &addr)
-                    .unwrap_or(0) as u64
-            } else {
-                token_opt
-                    .map(|token| token.balance_of(&recipient))
-                    .unwrap_or(0)
-            };
-            // Only backfill wallets that have NEVER received any SOV (balance == 0).
-            // Wallets with a positive balance already have their initial SOV (either from
-            // process_wallet_transactions, a previous backfill, or incoming transfers).
-            // Backfilling wallets that merely spent below initial_balance would inflate them.
-            if current_balance > 0 {
-                continue;
-            }
-            let deficit = wallet.initial_balance;
-            entries.push((wallet_key, deficit, wallet_id.clone()));
-        }
-        entries
-    }
-
-    /// Scan all blocks for duplicate TOKEN_BACKFILL_V1 TokenMint transactions and
-    /// correct any inflated Sled balances. Each restart that triggered the old
-    /// backfill code incorrectly minted an extra `initial_balance` worth of SOV.
-    /// This function detects duplicates and subtracts the excess.
-    ///
-    /// Safe to call on every startup — no-ops when store unavailable or no duplicates found.
-    pub fn repair_backfill_inflation(&self) -> usize {
-        let store = match self.get_store() {
-            Some(s) => s,
-            None => return 0,
-        };
-
-        let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-        let sov_storage_token_id = crate::storage::TokenId(sov_token_id);
-
-        // Track per-recipient mint amounts from TOKEN_BACKFILL_V1 mints (in block order).
-        let mut mint_history: std::collections::HashMap<[u8; 32], Vec<u128>> =
-            std::collections::HashMap::new();
-
-        for h in 0..=self.height {
-            let block = match store.get_block_by_height(h) {
-                Ok(Some(b)) => b,
-                _ => continue,
-            };
-            for tx in &block.transactions {
-                if tx.transaction_type != crate::types::transaction_type::TransactionType::TokenMint
-                {
-                    continue;
-                }
-                let is_backfill = std::str::from_utf8(&tx.memo)
-                    .map(|s| s.starts_with("TOKEN_BACKFILL_V1:"))
-                    .unwrap_or(false);
-                if !is_backfill {
-                    continue;
-                }
-                if let Some(mint_data) = tx.token_mint_data() {
-                    mint_history
-                        .entry(mint_data.to)
-                        .or_default()
-                        .push(mint_data.amount);
-                }
-            }
-        }
-
-        let mut corrections: Vec<(crate::storage::TokenId, crate::storage::Address, u128)> =
-            Vec::new();
-
-        for (wallet_key, amounts) in &mint_history {
-            if amounts.len() == 1 {
-                // Case 2: single TOKEN_BACKFILL_V1 mint — may be spurious if the wallet
-                // already had SOV from backfill_token_balances_from_contract (a direct Sled
-                // write, not a block transaction).  A partial top-up (amount < initial_balance)
-                // applied to a wallet that already had its full initial SOV causes:
-                //   current = initial_balance + mint_amount + subsequent_transfers
-                // which is inflation.  Correct by subtracting the spurious mint_amount.
-                // Idempotent: after correction current == initial_balance + subsequent_transfers,
-                // which is ≤ initial_balance only if no transfers happened, so the condition
-                // `current > initial_balance` won't trigger again on the next restart.
-                let mint_amount = amounts[0];
-                let wallet_id_hex = hex::encode(wallet_key);
-                let initial_balance = self
-                    .wallet_registry
-                    .get(&wallet_id_hex)
-                    .map(|w| w.initial_balance as u128)
-                    .unwrap_or(0);
-                if initial_balance > 0 && mint_amount < initial_balance {
-                    let addr = crate::storage::Address::new(*wallet_key);
-                    let current = store
-                        .get_token_balance(&sov_storage_token_id, &addr)
-                        .unwrap_or(0);
-                    if current > initial_balance {
-                        let corrected = current - mint_amount;
-                        info!(
-                            "🔧 Correcting spurious partial backfill for wallet {}: {} → {} \
-                             (removed spurious partial mint of {})",
-                            hex::encode(&wallet_key[..8]),
-                            current,
-                            corrected,
-                            mint_amount
-                        );
-                        corrections.push((sov_storage_token_id, addr, corrected));
-                    }
-                }
-                continue;
-            }
-            // Case 1: multiple TOKEN_BACKFILL_V1 mints — first is legitimate,
-            // all subsequent are duplicates from restarts before the fix.
-            let excess: u128 = amounts[1..].iter().sum();
-            let addr = crate::storage::Address::new(*wallet_key);
-            let current = store
-                .get_token_balance(&sov_storage_token_id, &addr)
-                .unwrap_or(0);
-            if current >= excess {
-                let corrected = current - excess;
-                info!(
-                    "🔧 Correcting backfill inflation for wallet {}: {} → {} \
-                     ({} duplicate mints, removing {} excess)",
-                    hex::encode(&wallet_key[..8]),
-                    current,
-                    corrected,
-                    amounts.len() - 1,
-                    excess
-                );
-                corrections.push((sov_storage_token_id, addr, corrected));
-            } else {
-                warn!(
-                    "⚠️ Cannot correct backfill inflation for wallet {}: \
-                     current {} < excess {}",
-                    hex::encode(&wallet_key[..8]),
-                    current,
-                    excess
-                );
-            }
-        }
-
-        let count = corrections.len();
-        if count > 0 {
-            match store.force_set_token_balances(&corrections) {
-                Ok(_) => info!("🔧 Repaired backfill inflation for {} wallets", count),
-                Err(e) => warn!("⚠️ Failed to write backfill corrections: {}", e),
-            }
-        }
-        count
-    }
-
-    /// Register a new wallet on the blockchain
-    pub fn register_wallet(
-        &mut self,
-        wallet_data: crate::transaction::WalletTransactionData,
-    ) -> Result<Hash> {
-        // Check if wallet already exists
-        let wallet_id_str = hex::encode(wallet_data.wallet_id.as_bytes());
-        if self.wallet_registry.contains_key(&wallet_id_str) {
-            return Err(anyhow::anyhow!(
-                "Wallet {} already exists on blockchain",
-                wallet_id_str
-            ));
-        }
-
-        // Create wallet registration transaction
-        let registration_tx = Transaction::new_wallet_registration(
-            wallet_data.clone(),
-            vec![], // Fee outputs handled separately
-            Signature {
-                signature: wallet_data.public_key.clone(),
-                public_key: PublicKey::new(wallet_data.public_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: wallet_data.created_at,
-            },
-            format!("Wallet registration for {}", wallet_data.wallet_name).into_bytes(),
-        );
-
-        // Add to pending transactions for inclusion in next block
-        // Wallet registration from node startup is a system operation - bypass signature validation
-        // This is consistent with how genesis funding directly inserts into wallet_registry
-        self.add_system_transaction(registration_tx.clone())?;
-
-        // Store in wallet registry immediately for queries
-        self.wallet_registry
-            .insert(wallet_id_str.clone(), wallet_data.clone());
-        self.wallet_blocks
-            .insert(wallet_id_str.clone(), self.height + 1);
-
-        // Mint SOV immediately in-memory so the balance is available regardless of whether
-        // the WalletRegistration tx ever lands in a block (e.g. when consensus is stalled).
-        // process_wallet_transactions() guards with `balance_of > 0` so it will skip
-        // wallets that are already credited, preventing double-minting on block commit.
-        if wallet_data.initial_balance > 0 {
-            let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-            self.ensure_sov_token_contract();
-            let mut wallet_id_bytes_arr = [0u8; 32];
-            wallet_id_bytes_arr.copy_from_slice(wallet_data.wallet_id.as_bytes());
-            let recipient_pk = Self::wallet_key_for_sov(&wallet_id_bytes_arr);
-            if let Some(token) = self.token_contracts.get_mut(&sov_token_id) {
-                if token.balance_of(&recipient_pk) == 0 {
-                    if let Err(e) = token.mint(&recipient_pk, wallet_data.initial_balance) {
-                        warn!(
-                            "register_wallet: failed to mint {} SOV for {}: {}",
-                            wallet_data.initial_balance,
-                            &wallet_id_str[..16.min(wallet_id_str.len())],
-                            e
-                        );
-                    } else {
-                        info!(
-                            "💰 register_wallet: minted {} SOV for wallet {} (in-memory)",
-                            wallet_data.initial_balance,
-                            &wallet_id_str[..16.min(wallet_id_str.len())]
-                        );
-                    }
-                }
-            }
-        }
-
-        Ok(registration_tx.hash())
-    }
-
-    /// Create a spendable UTXO for wallet funding (welcome bonus, migration, etc.)
-    ///
-    /// This creates an actual spendable output in the UTXO set, not just registry metadata.
-    /// The recipient is identified by their identity hash (32 bytes).
-    pub fn create_funding_utxo(
-        &mut self,
-        wallet_id: &str,
-        recipient_identity: &[u8],
-        amount: u64,
-    ) -> Hash {
-        let utxo_output = crate::transaction::TransactionOutput {
-            commitment: crate::types::hash::blake3_hash(
-                format!("funding_commitment_{}_{}", wallet_id, amount).as_bytes(),
-            ),
-            note: crate::types::hash::blake3_hash(format!("funding_note_{}", wallet_id).as_bytes()),
-            recipient: PublicKey::new(recipient_identity.to_vec()),
-        };
-        let utxo_hash = crate::types::hash::blake3_hash(
-            format!("funding_utxo:{}:{}", wallet_id, amount).as_bytes(),
-        );
-        self.utxo_set.insert(utxo_hash, utxo_output);
-        info!(
-            "💰 Created funding UTXO: {} SOV for wallet {}",
-            amount,
-            &wallet_id[..16.min(wallet_id.len())]
-        );
-        utxo_hash
-    }
-
-    /// Get wallet by ID
-    pub fn get_wallet(
-        &self,
-        wallet_id: &str,
-    ) -> Option<&crate::transaction::WalletTransactionData> {
-        self.wallet_registry.get(wallet_id)
-    }
-
-    /// Check if wallet exists
-    pub fn wallet_exists(&self, wallet_id: &str) -> bool {
-        self.wallet_registry.contains_key(wallet_id)
-    }
-
-    /// Get all wallets on the blockchain
-    pub fn list_all_wallets(&self) -> Vec<&crate::transaction::WalletTransactionData> {
-        self.wallet_registry.values().collect()
-    }
-
-    /// Get all wallets as HashMap
-    pub fn get_all_wallets(&self) -> &HashMap<String, crate::transaction::WalletTransactionData> {
-        &self.wallet_registry
-    }
-
-    /// Get wallet block confirmation count
-    pub fn get_wallet_confirmations(&self, wallet_id: &str) -> Option<u64> {
-        self.wallet_blocks.get(wallet_id).map(|block_height| {
-            if self.height >= *block_height {
-                self.height - block_height + 1
-            } else {
-                0
-            }
-        })
-    }
-
-    /// Get wallets for a specific owner identity
-    pub fn get_wallets_for_owner(
-        &self,
-        owner_identity_id: &Hash,
-    ) -> Vec<&crate::transaction::WalletTransactionData> {
-        self.wallet_registry
-            .values()
-            .filter(|wallet| wallet.owner_identity_id.as_ref() == Some(owner_identity_id))
-            .collect()
-    }
-
-    /// Process wallet transactions in a block
-    pub fn process_wallet_transactions(&mut self, block: &Block) -> Result<()> {
-        for transaction in &block.transactions {
-            if matches!(
-                transaction.transaction_type,
-                TransactionType::WalletRegistration | TransactionType::WalletUpdate
-            ) {
-                if let Some(wallet_data) = transaction.wallet_data() {
-                    let wallet_id_str = hex::encode(wallet_data.wallet_id.as_bytes());
-                    self.wallet_registry
-                        .insert(wallet_id_str.clone(), wallet_data.clone());
-                    self.wallet_blocks
-                        .insert(wallet_id_str.clone(), block.height());
-
-                    // Mint initial SOV balance for new wallets (block-authoritative).
-                    // This ensures the token contract is the source of truth and persists in the store.
-                    if transaction.transaction_type == TransactionType::WalletRegistration
-                        && wallet_data.initial_balance > 0
-                    {
-                        let sov_token_id = crate::contracts::utils::generate_lib_token_id();
-                        self.ensure_sov_token_contract();
-
-                        let mut wallet_id_bytes = [0u8; 32];
-                        wallet_id_bytes.copy_from_slice(wallet_data.wallet_id.as_bytes());
-                        let recipient_pk = Self::wallet_key_for_sov(&wallet_id_bytes);
-
-                        // Compute deficit: mint only what's missing to avoid double-counting
-                        // when register_wallet() already pre-minted in-memory.
-                        let current_balance = self
-                            .token_contracts
-                            .get(&sov_token_id)
-                            .map(|token| token.balance_of(&recipient_pk))
-                            .unwrap_or(0);
-                        let deficit = wallet_data
-                            .initial_balance
-                            .saturating_sub(current_balance);
-                        if deficit > 0 {
-                            if let Some(token) = self.token_contracts.get_mut(&sov_token_id) {
-                                if let Err(e) = token.mint(&recipient_pk, deficit) {
-                                    warn!(
-                                        "Failed to mint {} SOV for wallet {}: {}",
-                                        deficit,
-                                        &wallet_id_str[..16.min(wallet_id_str.len())],
-                                        e
-                                    );
-                                }
-                            }
-                        }
-                        // Always persist to sled on block commit so sled is authoritative,
-                        // even when balance was pre-minted in-memory by register_wallet().
-                        if let Some(store) = &self.store {
-                            if let Some(token) = self.token_contracts.get(&sov_token_id) {
-                                let store_ref: &dyn crate::storage::BlockchainStore =
-                                    store.as_ref();
-                                if let Err(e) = store_ref.put_token_contract(token) {
-                                    warn!(
-                                        "Failed to persist SOV token after wallet registration mint: {}",
-                                        e
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
     }
 
     // ========================================================================
@@ -4087,395 +2721,6 @@ impl Blockchain {
     ///
     /// If any two keys are identical the registration is rejected with an error
     /// describing which pair collides.  See the [`ValidatorInfo`] doc-comment for a
-    /// full description of each key's role and the security rationale for separation.
-    pub fn register_validator(&mut self, validator_info: ValidatorInfo) -> Result<Hash> {
-        // Check if validator already exists
-        if self
-            .validator_registry
-            .contains_key(&validator_info.identity_id)
-        {
-            return Err(anyhow::anyhow!(
-                "Validator {} already exists on blockchain",
-                validator_info.identity_id
-            ));
-        }
-
-        // Verify the identity exists
-        if !self
-            .identity_registry
-            .contains_key(&validator_info.identity_id)
-        {
-            return Err(anyhow::anyhow!(
-                "Identity {} must be registered before becoming a validator",
-                validator_info.identity_id
-            ));
-        }
-
-        // KEY SEPARATION ASSERTIONS
-        // Each key serves a distinct security domain; reuse collapses those boundaries.
-        if validator_info.consensus_key.is_empty() {
-            return Err(anyhow::anyhow!("Validator consensus_key must not be empty"));
-        }
-        if validator_info.networking_key.is_empty() {
-            return Err(anyhow::anyhow!(
-                "Validator networking_key must not be empty"
-            ));
-        }
-        if validator_info.rewards_key.is_empty() {
-            return Err(anyhow::anyhow!("Validator rewards_key must not be empty"));
-        }
-        if validator_info.consensus_key == validator_info.networking_key {
-            return Err(anyhow::anyhow!(
-                "Validator key separation violation: consensus_key and networking_key must be different keys. \
-                 Reusing the same key across roles collapses security domain boundaries."
-            ));
-        }
-        if validator_info.consensus_key == validator_info.rewards_key {
-            return Err(anyhow::anyhow!(
-                "Validator key separation violation: consensus_key and rewards_key must be different keys. \
-                 A compromised consensus key must not give an attacker control over staking rewards."
-            ));
-        }
-        if validator_info.networking_key == validator_info.rewards_key {
-            return Err(anyhow::anyhow!(
-                "Validator key separation violation: networking_key and rewards_key must be different keys. \
-                 A compromised network identity key must not give an attacker access to reward funds."
-            ));
-        }
-
-        // SECURITY: Validate minimum requirements for validator eligibility
-        // Edge nodes (minimal storage, no consensus capability) cannot become validators
-        // Genesis bootstrap: Allow 1,000 SOV minimum for initial validator setup
-        // Production: Require 100,000 SOV minimum after genesis
-        let min_stake = if self.height == 0 { 1_000 } else { 100_000 };
-        if validator_info.stake < min_stake {
-            return Err(anyhow::anyhow!(
-                "Insufficient stake for validator: {} SOV (minimum: {} SOV required)",
-                validator_info.stake,
-                min_stake
-            ));
-        }
-
-        // Storage requirement: Only enforce for production validators after genesis
-        // Genesis validators (height 0) can register with any storage amount for testing
-        if self.height > 0 && validator_info.storage_provided < 10_737_418_240 {
-            // 10 GB in bytes
-            return Err(anyhow::anyhow!(
-                "Insufficient storage for validator: {} bytes (minimum: 10 GB required for blockchain storage)",
-                validator_info.storage_provided
-            ));
-        }
-
-        // Create validator registration transaction (using Identity type as placeholder until we add Validator type)
-        let validator_tx_data = IdentityTransactionData {
-            did: validator_info.identity_id.clone(),
-            display_name: format!("Validator: {}", validator_info.network_address),
-            public_key: validator_info.consensus_key.clone(),
-            ownership_proof: vec![], // Empty for system validator registration
-            identity_type: "validator".to_string(),
-            did_document_hash: crate::types::hash::blake3_hash(
-                format!(
-                    "validator:{}:{}",
-                    validator_info.identity_id, validator_info.registered_at
-                )
-                .as_bytes(),
-            ),
-            created_at: validator_info.registered_at,
-            registration_fee: 0, // No fee for validator registration (paid via stake)
-            dao_fee: 0,
-            controlled_nodes: Vec::new(),
-            owned_wallets: Vec::new(),
-        };
-
-        let registration_tx = Transaction::new_identity_registration(
-            validator_tx_data,
-            vec![], // Fee outputs handled separately
-            Signature {
-                signature: validator_info.consensus_key.clone(),
-                public_key: PublicKey::new(validator_info.consensus_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: validator_info.registered_at,
-            },
-            format!(
-                "Validator registration for {} with stake {}",
-                validator_info.identity_id, validator_info.stake
-            )
-            .into_bytes(),
-        );
-
-        // Add to pending transactions for inclusion in next block
-        self.add_pending_transaction(registration_tx.clone())?;
-
-        // Store in validator registry immediately for queries
-        self.validator_registry
-            .insert(validator_info.identity_id.clone(), validator_info.clone());
-        self.validator_blocks
-            .insert(validator_info.identity_id.clone(), self.height + 1);
-
-        info!(
-            " Validator {} registered with {} SOV stake and {} bytes storage",
-            validator_info.identity_id, validator_info.stake, validator_info.storage_provided
-        );
-
-        Ok(registration_tx.hash())
-    }
-
-    /// Get validator by identity ID
-    pub fn get_validator(&self, identity_id: &str) -> Option<&ValidatorInfo> {
-        self.validator_registry.get(identity_id)
-    }
-
-    /// Check if validator exists
-    pub fn validator_exists(&self, identity_id: &str) -> bool {
-        self.validator_registry.contains_key(identity_id)
-    }
-
-    /// Get all validators on the blockchain
-    pub fn list_all_validators(&self) -> Vec<&ValidatorInfo> {
-        self.validator_registry.values().collect()
-    }
-
-    /// Get all active validators
-    pub fn get_active_validators(&self) -> Vec<&ValidatorInfo> {
-        self.validator_registry
-            .values()
-            .filter(|v| v.status == "active")
-            .collect()
-    }
-
-    /// Get all validators as HashMap
-    pub fn get_all_validators(&self) -> &HashMap<String, ValidatorInfo> {
-        &self.validator_registry
-    }
-
-    /// Update validator information
-    pub fn update_validator(
-        &mut self,
-        identity_id: &str,
-        updated_info: ValidatorInfo,
-    ) -> Result<Hash> {
-        // Check if validator exists
-        if !self.validator_registry.contains_key(identity_id) {
-            return Err(anyhow::anyhow!(
-                "Validator {} not found on blockchain",
-                identity_id
-            ));
-        }
-
-        // Create update transaction
-        let validator_tx_data = IdentityTransactionData {
-            did: updated_info.identity_id.clone(),
-            display_name: format!("Validator Update: {}", updated_info.network_address),
-            public_key: updated_info.consensus_key.clone(),
-            ownership_proof: vec![],
-            identity_type: "validator".to_string(),
-            did_document_hash: crate::types::hash::blake3_hash(
-                format!(
-                    "validator_update:{}:{}",
-                    updated_info.identity_id, updated_info.last_activity
-                )
-                .as_bytes(),
-            ),
-            created_at: updated_info.last_activity,
-            registration_fee: 0,
-            dao_fee: 0,
-            controlled_nodes: Vec::new(),
-            owned_wallets: Vec::new(),
-        };
-
-        let update_tx = Transaction::new_identity_update(
-            validator_tx_data,
-            vec![],
-            vec![],
-            100, // Update fee
-            Signature {
-                signature: updated_info.consensus_key.clone(),
-                public_key: PublicKey::new(updated_info.consensus_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: updated_info.last_activity,
-            },
-            format!("Validator update for {}", identity_id).into_bytes(),
-        );
-
-        // Add to pending transactions
-        self.add_pending_transaction(update_tx.clone())?;
-
-        // Update registry
-        self.validator_registry
-            .insert(identity_id.to_string(), updated_info);
-
-        Ok(update_tx.hash())
-    }
-
-    /// Unregister a validator
-    pub fn unregister_validator(&mut self, identity_id: &str) -> Result<Hash> {
-        // Check if validator exists
-        if !self.validator_registry.contains_key(identity_id) {
-            return Err(anyhow::anyhow!(
-                "Validator {} not found on blockchain",
-                identity_id
-            ));
-        }
-
-        // Get validator info
-        let mut validator_info = self.validator_registry.get(identity_id).unwrap().clone();
-        validator_info.status = "inactive".to_string();
-
-        // Create unregistration transaction
-        let unregister_tx = Transaction::new_identity_revocation(
-            identity_id.to_string(),
-            vec![],
-            100,
-            Signature {
-                signature: validator_info.consensus_key.clone(),
-                public_key: PublicKey::new(validator_info.consensus_key.clone()),
-                algorithm: SignatureAlgorithm::Dilithium2,
-                timestamp: validator_info.last_activity,
-            },
-            format!("Validator unregistration for {}", identity_id).into_bytes(),
-        );
-
-        // Add to pending transactions
-        self.add_pending_transaction(unregister_tx.clone())?;
-
-        // Update status in registry
-        self.validator_registry
-            .insert(identity_id.to_string(), validator_info);
-
-        info!("Validator {} unregistered", identity_id);
-
-        Ok(unregister_tx.hash())
-    }
-
-    /// Get validator block confirmation count
-    pub fn get_validator_confirmations(&self, identity_id: &str) -> Option<u64> {
-        self.validator_blocks.get(identity_id).map(|block_height| {
-            if self.height >= *block_height {
-                self.height - block_height + 1
-            } else {
-                0
-            }
-        })
-    }
-
-    // ========================================================================
-    // VALIDATOR SYNCHRONIZATION (Issue #5)
-    // ========================================================================
-
-    /// Get active validator set as (identity_id, stake) tuples for consensus integration
-    ///
-    /// This is the primary interface for consensus to query the validator set.
-    /// Used to keep consensus and blockchain validator registries in sync.
-    /// Only returns validators with active status and non-zero stake.
-    pub fn get_active_validator_set_for_consensus(&self) -> Vec<(String, u64)> {
-        self.get_active_validators()
-            .iter()
-            .map(|v| (v.identity_id.clone(), v.stake))
-            .collect()
-    }
-
-    /// Get total stake of all active validators
-    pub fn get_total_validator_stake(&self) -> u64 {
-        self.get_active_validators()
-            .iter()
-            .fold(0u64, |sum, v| sum.saturating_add(v.stake))
-    }
-
-    /// Check if a validator is in good standing (active status and sufficient stake)
-    pub fn is_validator_active(&self, identity_id: &str) -> bool {
-        if let Some(validator) = self.validator_registry.get(identity_id) {
-            // Validator must have active status and non-zero stake
-            validator.status == "active" && validator.stake > 0
-        } else {
-            false
-        }
-    }
-
-    /// Emit validator set changed event for consensus integration
-    /// Call this whenever the validator set changes to keep consensus in sync
-    pub fn sync_validator_set_to_consensus(&self) {
-        let active_validators = self.get_active_validators();
-        info!(
-            "Validator set sync: {} active validators with {} total stake",
-            active_validators.len(),
-            self.get_total_validator_stake()
-        );
-
-        // In production, this would emit an event that consensus subscribes to
-        // For now, log the validator set for audit trail
-        for validator in active_validators {
-            debug!(
-                "Validator in sync: {} (stake: {}, joined at height: {})",
-                validator.identity_id, validator.stake, validator.registered_at
-            );
-        }
-    }
-
-    /// Process ValidatorRegistration transactions from a newly-committed block.
-    ///
-    /// This mirrors the replay logic in `load_from_file` so that validator state is
-    /// updated in real-time as blocks are mined, not only on restart.
-    pub fn process_validator_registration_transactions(&mut self, block: &Block) {
-        let height = block.height();
-        for tx in &block.transactions {
-            if let Some(validator_data) = tx.validator_data() {
-                let status = match validator_data.operation {
-                    crate::transaction::ValidatorOperation::Register => "active",
-                    crate::transaction::ValidatorOperation::Update => "active",
-                    crate::transaction::ValidatorOperation::Unregister => "inactive",
-                };
-                let validator_info = ValidatorInfo {
-                    identity_id: validator_data.identity_id.clone(),
-                    stake: validator_data.stake,
-                    storage_provided: validator_data.storage_provided,
-                    consensus_key: validator_data.consensus_key.clone(),
-                    networking_key: validator_data.networking_key.clone(),
-                    rewards_key: validator_data.rewards_key.clone(),
-                    network_address: validator_data.network_address.clone(),
-                    commission_rate: validator_data.commission_rate,
-                    status: status.to_string(),
-                    registered_at: height,
-                    last_activity: height,
-                    blocks_validated: 0,
-                    slash_count: 0,
-                    admission_source: ADMISSION_SOURCE_ONCHAIN_GOVERNANCE.to_string(),
-                    governance_proposal_id: None,
-                    oracle_key_id: None,
-                };
-                self.validator_registry
-                    .insert(validator_data.identity_id.clone(), validator_info);
-                self.validator_blocks
-                    .insert(validator_data.identity_id.clone(), height);
-                info!(
-                    "Registered new validator {} with {} SOV stake",
-                    &validator_data.identity_id[..validator_data.identity_id.len().min(40)],
-                    validator_data.stake
-                );
-            }
-        }
-    }
-
-    /// Process validator transactions in a block
-    pub fn process_validator_transactions(&mut self, block: &Block) -> Result<()> {
-        for transaction in &block.transactions {
-            if let Some(identity_data) = transaction.identity_data() {
-                if identity_data.identity_type == "validator" {
-                    // Extract validator info from identity transaction
-                    // This is a simplified version - in production, you'd have a dedicated ValidatorTransactionData
-                    if let Some(validator_info) = self.validator_registry.get(&identity_data.did) {
-                        let mut updated_info = validator_info.clone();
-                        updated_info.last_activity = identity_data.created_at;
-                        updated_info.blocks_validated += 1;
-
-                        self.validator_registry
-                            .insert(identity_data.did.clone(), updated_info);
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
     /// Process contract deployment and execution transactions from a block
     pub fn process_contract_transactions(&mut self, block: &Block) -> Result<()> {
         for transaction in &block.transactions {
